requirements.txt
appdirs==1.4.4
APScheduler==3.10.4
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
cssselect==1.2.0
Django==5.0.3
djangorestframework==3.15.2
fake-useragent==1.5.1
h11==0.14.0
idna==3.6
importlib_metadata==7.1.0
kiwipiepy==0.18.0
kiwipiepy_model==0.18.0
lxml==5.2.1
lxml_html_clean==0.1.1
mysqlclient==2.2.4
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.1
parse==1.20.1
pyee==11.1.0
pyppeteer==2.0.0
pyquery==2.0.0
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-html @ git+https://github.com/psf/requests-html.git@075ac162dc62fc532037df0d98954ab840a97516
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.2
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzlocal==5.2
urllib3==1.26.18
w3lib==2.1.2
webdriver-manager==4.0.2
websockets==10.4
wsproto==1.2.0
zipp==3.18.1
